move worker to separate file
